Add support for GITHUB_TOKEN env var for auth
`gh auth login` will not create a ~/.config/gh/hosts.yml file if it
detects a GITHUB_TOKEN env var (though one can force gh to do that), and
a missing config file will fail. Instead of forcing it, this commit
allows the GITHUB_TOKEN env var to be accepted.